requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.8.0
transformers>=4.40.0
torch>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
groq>=0.4.0 